move equivalance class to its own file
